requirements.txt
amqp==2.5.2
astroid==2.3.3
autopep8==1.4.4
billiard==3.6.1.0
celery==4.3.0
certifi==2019.11.28
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.7
django-celery-beat==1.5.0
django-celerybeat-status==0.0.10
django-timezone-field==3.1
gunicorn==20.0.2
honcho==1.0.1
idna==2.8
importlib-metadata==0.23
isort==4.3.21
kombu==4.6.6
kombu-sqlalchemy==1.1.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==7.2.0
psycopg2-binary==2.8.4
pycodestyle==2.5.0
pylint==2.4.4
python-crontab==2.4.0
python-dateutil==2.8.1
python-decouple==3.3
pytz==2019.3
redis==3.3.11
requests==2.22.0
six==1.13.0
SQLAlchemy==1.3.11
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.7
vine==1.3.0
whitenoise==4.1.4
wrapt==1.11.2
zipp==0.6.0
